+++ src/components/NotFoundPage/NotFoundContent.style.ts
@@ -1,0 +1,17 @@
+const styles = {
+  title: {
+    marginBottom: '2rem',
+  },
+  button: {
+    marginBottom: '3rem',
+  },
+  carouselContainer: {
+    width: '100%',
+    maxWidth: '800px',
+  },
+  carouselTitle: {
+    marginBottom: '1rem',
+  },
+};
+
+export default styles;
