+++ src/pages/NotFoundPage/NotFound.tsx
@@ -1,14 +1,7 @@
-import { Typography, Button, Box } from '@mui/material';
-import { useNavigate } from 'react-router-dom';
-import BestBookCarousel from '@components/BookSearchPage/BestBookCarousel';
+import NotFoundContent from '@components/NotFoundPage/NotFoundContent';
+import { Box } from '@mui/material';
 
 const NotFoundPage = (): JSX.Element => {
-  const navigate = useNavigate();
-
-  const handleGoHome = () => {
-    navigate('/');
-  };
-
   return (
     <Box
       sx={{
@@ -21,23 +14,7 @@ const NotFoundPage = (): JSX.Element => {
         backgroundColor: '#ffffff',
       }}
     >
-      <Typography variant="h4" sx={{ marginBottom: '2rem' }}>
-        페이지를 찾을 수 없습니다
-      </Typography>
-      <Button
-        variant="contained"
-        color="primary"
-        onClick={handleGoHome}
-        sx={{ marginBottom: '3rem' }}
-      >
-        메인 페이지로 이동
-      </Button>
-      <Box sx={{ width: '100%', maxWidth: '800px' }}>
-        <Typography variant="h5" sx={{ marginBottom: '1rem' }}>
-          베스트셀러
-        </Typography>
-        <BestBookCarousel />
-      </Box>
+      <NotFoundContent />
     </Box>
   );
 };
